Update the tooltip eslint rule as we now export the new one from main bundle and the old one from deprecated
